Rename to avoid use of blueprint.blueprint in the Python API.
